intersection detection: publish detected data

See merge request ostfaliacup/af24/gruppe2!18

+++ src/decider/States/Approaching_Crossing.h
@@ -26,6 +26,6 @@ class Approaching_Crossing: State {
         Approaching_Crossing(){}
         Approaching_Crossing(const Approaching_Crossing& other);
         Approaching_Crossing& operator=(const Approaching_Crossing& other);
-        double* smooth_speed(int current_speed);
+        double* smooth_speed(int16_t current_speed);
 };
